+++ src/Controllers/ProviderController.php
@@ -1,20 +1,48 @@
 <?php
     require_once(__DIR__.'/../../autoload.php');
     require_once('src/helpers/SessionValidate.php');
+    require_once('src/helpers/Validate.php');
 
     class ProviderController{
         public function index(){
-            $provider = new AddProviderView();
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $provider = new AddProviderView();
+            }else{
+                return redirect('/provider', 'Usuário sem permissão.');
+            }
         }
 
         public function store($params){
-            $provider = new ProviderModel();
-            $result = $provider->create($params);
-
-            if($result){
-                return redirect('/add/provider', 'Forncedor cadastrado com sucesso');
+            $data = array(
+                'Nome' => $params['nome'],
+                'CNPJ' => $params['cnpj'],
+                'Endereco' => $params['endereco'],
+                'Telefone' => $params['telefone'],
+                'Email' => $params['email']
+            );
+            $validator = array(
+                'Nome' => 'required',
+                'CNPJ' => 'required|cnpj',
+                'Endereco' => 'required',
+                'Telefone' => 'required',
+                'Email' => 'required|email',
+            );
+            $resultValidation = validate($data,$validator);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                if($resultValidation['success']){
+                    $provider = new ProviderModel();
+                    $result = $provider->create($params);
+    
+                    if($result){
+                        return redirect('/provider', 'Forncedor cadastrado com sucesso');
+                    }else{
+                        return redirect('/add/provider', 'Ocorreu um erro interno. Contate os desenvolvedores.');
+                    }
+                }else{
+                    return redirect('/provider', $resultValidation['message']);
+                }
             }else{
-                return redirect('/add/provider', 'Algo deu errado, tente novamente');
+                return redirect('/provider', 'Usuário sem permissão.');
             }
         }
 
@@ -29,31 +57,85 @@
             echo json_encode($provider->getProviderById($params));
         }
 
-        public function edit($params){
+        public function getProviderByName($params){
             $provider = new ProviderModel();
-            $params['provider'] = $provider->getProviderById($params);
-            $provider = new EditProviderView($params);
+            echo json_encode($provider->getProviderByName($params));
+        }
+
+        public function edit($params){
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                $provider = new ProviderModel();
+                $params['provider'] = $provider->getProviderById($params);
+                $provider = new EditProviderView($params);
+             }else{
+                return redirect('/provider', 'Usuário sem permissão.');
+            }
         }
 
         public function update($params){
-            $provider = new ProviderModel();
-            $result = $provider->update($params);
+            $data = array(
+                'Nome' => $params['nome'],
+                'CNPJ' => $params['cnpj'],
+                'Endereco' => $params['endereco'],
+                'Telefone' => $params['telefone'],
+                'Email' => $params['email'],
+                'ID_do_fornecedor' => $params['providerId']
+            );
+            $validator = array(
+                'Nome' => 'required',
+                'CNPJ' => 'required|cnpj',
+                'Endereco' => 'required',
+                'Telefone' => 'required',
+                'Email' => 'required|email',
+                'ID_do_fornecedor' => 'required'
+            );
+            $resultValidation = validate($data,$validator);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                if($resultValidation['success']){
+                    $provider = new ProviderModel();
+                    $result = $provider->update($params);
 
-            if($result){
-                return redirect('/provider', 'Fornecedor atualizado com sucesso');
+                    if($result){
+                        return redirect('/provider', 'Fornecedor atualizado com sucesso');
+                    }else{
+                        return redirect('/provider', 'Ocorreu um erro interno. Contate os desenvolvedores.');
+                    }
+                }else{
+                    return redirect('/provider', $resultValidation['message']);
+                }
             }else{
-                return redirect('/provider', 'Algo deu errado, tente novamente');
+                return redirect('/provider', 'Usuário sem permissão.');
             }
         }
         public function delete($params){
-            $provider = new ProviderModel();
-            $result = $provider->delete($params);
+            $data = array(
+                'ID_do_fornecedor' => $params['id']
+            );
+            $validator = array(
+                'ID_do_fornecedor' => 'required'
+            );
+            $resultValidation = validate($data,$validator);
+            if($_SESSION['dados_usuario']['nivelAcesso'] == 2){
+                if($resultValidation['success']){
+                    $provider = new ProviderModel();
+                    $result = $provider->delete($params);
 
-            if($result){
-                return redirect('/provider', 'Fornecedor deletado com sucesso');
+                    if($result){
+                        return redirect('/provider', 'Fornecedor deletado com sucesso');
+                    }else{
+                        return redirect('/provider', 'Ocorreu um erro interno. Contate os desenvolvedores.');
+                    }
+                }else{
+                    return redirect('/provider', $resultValidation['message']);
+                }
+
             }else{
-                return redirect('/provider', 'Algo deu errado, tente novamente');
+                return redirect('/provider', 'Usuário sem permissão.');
             }
+        }
+        public function getAllProviders(){
+            $provider = new ProviderModel();
+            echo json_encode($provider->getAllProviders());
         }
     }
 ?>
